+++ spec/RequestSpec.php
@@ -5,8 +5,8 @@ use ochenta\Request;
 describe('Request', function() {
     describe('->__construct', function() {
         it('assigns method in uppercase', function() {
-            expect((new Request('get', '/'))->getMethod())->toBe('GET');
-            expect((new Request('gET', '/'))->getMethod())->toBe('GET');
+            expect((new Request('get', '/', ['Host' => 'example.com']))->getMethod())->toBe('GET');
+            expect((new Request('gET', '/', ['Host' => 'example.com']))->getMethod())->toBe('GET');
         });
 
         it('assigns uri from parse_url array', function() {
@@ -42,16 +42,23 @@ describe('Request', function() {
         });
 
         it('assigns body, defaults to null', function() {
-            expect((new Request('GET', '/'))->getBody())->toBeNull();
+            expect((new Request('GET', '/', ['Host' => 'example.com']))->getBody())->toBeNull();
         });
 
         it('assigns body as resource if given is scalar', function() {
-           expect((new Request('GET', '/', [], 'Hello World'))->getBody())->toBeA('resource');
+           expect((new Request('GET', '/', ['Host' => 'example.com'], 'Hello World'))->getBody())->toBeA('resource');
+        });
+
+        it('throws InvalidArgumentException on missing host header', function() {
+            expect(function() {
+                new Request('GET', '/');
+            })
+            ->toThrow(new InvalidArgumentException);
         });
 
         it('throws InvalidArgumentException on invalid body', function() {
             expect(function() {
-                new Request('GET', '/', [], []);
+                new Request('GET', '/', ['Host' => 'example.com'], []);
             })
             ->toThrow(new InvalidArgumentException);
         });
@@ -59,7 +66,7 @@ describe('Request', function() {
 
     describe('->getTarget', function() {
         it('return forward slash when path component of the uri is not defined', function() {
-            expect((new Request('GET', '?queryString'))->getTarget())->toBe('/?queryString');
+            expect((new Request('GET', '?queryString', ['Host' => 'example.com']))->getTarget())->toBe('/?queryString');
         });
 
         it('return path and query string of the uri', function() {
@@ -69,43 +76,71 @@ describe('Request', function() {
 
     describe('->getMediaType', function() {
         it('returns null when Content-Type is not defined', function() {
-            expect((new Request('GET', '/'))->getMediaType())->toBeNull();
+            expect((new Request('GET', '/', ['Host' => 'example.com']))->getMediaType())->toBeNull();
         });
 
         it('returns Content-Type header in lowercase', function() {
-            expect((new Request('GET', '/', ['Content-Type' => ['TEXT/PLAIN']]))->getMediaType())->toBe('text/plain');
+            $req = new Request('GET', '/', [
+                'Host' => 'example.com',
+                'Content-Type' => ['TEXT/PLAIN']
+            ]);
+            expect($req->getMediaType())->toBe('text/plain');
         });
 
         it('returns Content-Type header without charset', function() {
-            expect((new Request('GET', '/', ['Content-Type' => ['text/plain;charset=ISO-8859-4']]))->getMediaType())->toBe('text/plain');
+            $req = new Request('GET', '/', [
+                'Host' => 'example.com',
+                'Content-Type' => ['text/plain;charset=ISO-8859-4']
+            ]);
+            expect($req->getMediaType())->toBe('text/plain');
         });
     });
 
     describe('->getCharset', function() {
         it('returns null when Content-Type is not defined', function() {
-            expect((new Request('GET', '/'))->getCharset())->toBeNull();
+            expect((new Request('GET', '/', ['Host' => 'example.com']))->getCharset())->toBeNull();
         });
 
         it('returns null when Content-Type charset is not given', function() {
-            expect((new Request('GET', '/', ['Content-Type' => ['text/plain;important=yes']]))->getCharset())->toBeNull();
+            $req = new Request('GET', '/', [
+                'Host' => 'example.com',
+                'Content-Type' => ['text/plain;important=yes']
+            ]);
+            expect($req->getCharset())->toBeNull();
         });
 
         it('returns normalized Content-Type charset', function() {
-            expect((new Request('GET', '/', ['Content-Type' => ['text/plain;CHARSET="UTF-8"']]))->getCharset())->toBe('utf-8');
+            $req = new Request('GET', '/', [
+                'Host' => 'example.com',
+                'Content-Type' => ['text/plain;CHARSET="UTF-8"']
+            ]);
+            expect($req->getCharset())->toBe('utf-8');
         });
     });
 
     describe('->isForm', function() {
         it('returns true when Content-Type is application/x-www-form-urlencoded', function() {
-            expect((new Request('GET', '/', ['Content-Type' => ['application/x-www-form-urlencoded']]))->isForm())->toBe(TRUE);
+            $req = new Request('GET', '/', [
+                'Host' => 'example.com',
+                'Content-Type' => ['application/x-www-form-urlencoded']
+            ]);
+            expect($req->isForm())->toBe(TRUE);
         });
 
         it('returns true when Content-Type is multipart/form-data', function() {
-            expect((new Request('GET', '/', ['Content-Type' => ['multipart/form-data']]))->isForm())->toBe(TRUE);
+            $req = new Request('GET', '/', [
+                'Host' => 'example.com',
+                'Content-Type' => ['multipart/form-data']
+            ]);
+            expect($req->isForm())->toBe(TRUE);
         });
 
         it('returns false when Content-Type is text/plain', function() {
-            expect((new Request('GET', '/', ['Content-Type' => ['text/plain']]))->isForm())->toBe(FALSE);
+            $req = new Request('GET', '/', [
+                'Host' => 'example.com',
+                'Content-Type' => ['text/plain']
+            ]);
+            expect($req->isForm())->toBe(FALSE);
         });
     });
 });